채팅방 data Layer 계층 분리
